♻️ Moved cell summary table generation to main module
Moved calculate_metrics_per_cell function from R report script to main python module, as it's a main functionality/output that belongs to the core processing of SQANTI-sc, not exclusively to the report. The function has been translated to python and optimized to reduce code space. Changes to the R script involve removing the function and renaming some features to match new format from main code.